pyproject.toml
[tool.poetry]
name = "steward"
version = "0.1.0"
description = ""
authors = ["Giannis Papaioannou <giannispapcod7@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
rich = "^9.1.0"
typer = "^0.3.2"
python-decouple = "^3.3"
influxdb = "^5.3.0"
speedtest-cli = "^2.1.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
